compile err fix

+++ libft/ft_putchar_fd.c
@@ -6,13 +6,13 @@
 /*   By: jkoupy <jkoupy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/09 14:14:53 by jkoupy            #+#    #+#             */
-/*   Updated: 2023/09/24 12:04:47 by jkoupy           ###   ########.fr       */
+/*   Updated: 2023/09/24 16:04:56 by jkoupy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_putchar_fd(char c, int fd)
+int	ft_putchar_fd(char c, int fd)
 {
 	write(fd, &c, 1);
 	return (1);
