+++ libft/ft_putstr_fd.c
@@ -6,16 +6,16 @@
 /*   By: jkoupy <jkoupy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/09 14:24:38 by jkoupy            #+#    #+#             */
-/*   Updated: 2023/09/24 14:13:06 by jkoupy           ###   ########.fr       */
+/*   Updated: 2023/09/24 16:06:10 by jkoupy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 #include <stdio.h>
 
-size_t	ft_putstr_fd(char *s, int fd)
+int	ft_putstr_fd(char *s, int fd)
 {
-	size_t	i;
+	int	i;
 
 	if (!s)
 		return (ft_putstr_fd("(null)", fd));
